+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailFragment.kt
@@ -65,6 +65,7 @@ import im.vector.matrix.android.api.session.room.send.SendState
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.room.timeline.getLastMessageContent
 import im.vector.matrix.android.api.session.room.timeline.getTextEditableContent
+import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.user.model.User
 import im.vector.riotx.R
 import im.vector.riotx.core.di.ScreenComponent
@@ -246,6 +247,14 @@ class RoomDetailFragment :
                 is SendMode.QUOTE -> enterSpecialMode(mode.timelineEvent, R.drawable.ic_quote, false)
                 is SendMode.REPLY -> enterSpecialMode(mode.timelineEvent, R.drawable.ic_reply, false)
             }
+        }
+
+        roomDetailViewModel.selectSubscribe(RoomDetailViewState::syncState) { syncState ->
+            syncProgressBar.visibility = when (syncState) {
+                is SyncState.RUNNING -> if (syncState.afterPause) View.VISIBLE else View.GONE
+                else                 -> View.GONE
+            }
+            syncProgressBarWrap.visibility = syncProgressBar.visibility
         }
     }