+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailViewState.kt
@@ -21,9 +21,9 @@ import com.airbnb.mvrx.MvRxState
 import com.airbnb.mvrx.Uninitialized
 import im.vector.matrix.android.api.session.events.model.Event
 import im.vector.matrix.android.api.session.room.model.RoomSummary
-import im.vector.matrix.android.api.session.room.model.tombstone.RoomTombstoneContent
 import im.vector.matrix.android.api.session.room.timeline.Timeline
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
+import im.vector.matrix.android.api.session.sync.SyncState
 import im.vector.matrix.android.api.session.user.model.User
 
 /**
@@ -50,7 +50,8 @@ data class RoomDetailViewState(
         val sendMode: SendMode = SendMode.REGULAR,
         val isEncrypted: Boolean = false,
         val tombstoneEvent: Event? = null,
-        val tombstoneEventHandling: Async<String> = Uninitialized
+        val tombstoneEventHandling: Async<String> = Uninitialized,
+        val syncState: SyncState = SyncState.IDLE
 ) : MvRxState {
 
     constructor(args: RoomDetailArgs) : this(roomId = args.roomId, eventId = args.eventId)